Separacion de mains + python de correl + modificaciones en general.c
general.c : guille nos dijo que agreguemos algunos casteos de double.

Separacion de mains: Ahora hay un  main_1a para el ejercicio 1a y otro main_1b para el 1b.  La gracia es no cambiarlo constantemente y perderlo todas las veces.

En los mains se agregó además comentarios sobre lo que hacen al principio.

Correlacion(Tau).ipynb : es un archivo de python que le das el .txt que sale de main_1a y te permite estudiar termalizacion y correlacion. Los datos que obtengas de este analisis los necesitas para correr en main_1b.

+++ src/general.c
@@ -15,7 +15,7 @@ double gaussiana(double mu, double sigma)
 	{
 		z += random();
 	}
-	z = sqrt(12*n)*(z/n - 0.5);
+	z = sqrt(12*(double)n)*(z/(double)n - 0.5);
 	return z*sigma+mu;
 }
